+++ sdv/constraints/base.py
@@ -11,10 +11,10 @@ from copulas.univariate import GaussianUnivariate
 from rdt import HyperTransformer
 from rdt.transformers import BinaryEncoder, FloatFormatter, OneHotEncoder, UnixTimestampEncoder
 
+from sdv._utils import _format_invalid_values_string, _groupby_list
 from sdv.constraints.errors import (
     AggregateConstraintsError, ConstraintMetadataError, MissingConstraintColumnError)
 from sdv.errors import ConstraintsNotMetError
-from sdv.utils import _format_invalid_values_string, _groupby_list
 
 LOGGER = logging.getLogger(__name__)
