+++ sdv/metadata/single_table.py
@@ -11,6 +11,9 @@ from itertools import combinations
 import pandas as pd
 from rdt.transformers.pii.anonymization import SDTYPE_ANONYMIZERS, is_faker_function
 
+from sdv._utils import (
+    _cast_to_iterable, _format_invalid_values_string, _get_datetime_format, _is_boolean_type,
+    _is_datetime_type, _is_numerical_type, _load_data_from_csv, _validate_datetime_format)
 from sdv.errors import InvalidDataError
 from sdv.metadata.errors import InvalidMetadataError
 from sdv.metadata.metadata_upgrader import convert_metadata
@@ -18,9 +21,6 @@ from sdv.metadata.utils import read_json, validate_file_does_not_exist
 from sdv.metadata.validation import validate_address_sdtypes, validate_gps_sdtypes
 from sdv.metadata.visualization import (
     create_columns_node, create_summarized_columns_node, visualize_graph)
-from sdv.utils import (
-    _cast_to_iterable, _format_invalid_values_string, _get_datetime_format, _is_boolean_type,
-    _is_datetime_type, _is_numerical_type, _load_data_from_csv, _validate_datetime_format)
 
 LOGGER = logging.getLogger(__name__)
